Rename `features/Home` -> `features/EntryList`
