+++ src/features/EntryList/PickedEntryList.tsx
